+++ example/lib/pages/welcome_page.dart
@@ -7,6 +7,7 @@ import 'package:better_player_example/pages/cache_page.dart';
 import 'package:better_player_example/pages/controller_controls_page.dart';
 import 'package:better_player_example/pages/controls_configuration_page.dart';
 import 'package:better_player_example/pages/event_listener_page.dart';
+import 'package:better_player_example/pages/fade_placeholder_page.dart';
 import 'package:better_player_example/pages/hls_subtitles_page.dart';
 import 'package:better_player_example/pages/hls_tracks_page.dart';
 import 'package:better_player_example/pages/memory_player_page.dart';
@@ -123,6 +124,9 @@ class _WelcomePageState extends State<WelcomePage> {
       }),
       _buildExampleElementWidget("Reusable video list page", () {
         _navigateToPage(ReusableVideoListPage());
+      }),
+      _buildExampleElementWidget("Fade placeholder page", () {
+        _navigateToPage(FadePlaceholderPage());
       }),
     ];
   }