+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Fixed Android native build warnings
 * Fixed placeholder until play issues
 * Lint fixes
-
+* Added placeholderOnTop to the BetterPlayerConfiguration
 
 ## 0.0.42
 * Fixed resolution issue